Fix negative temperatures and stash raw packet
This fixes decoding negative (celsius) temperatures so that we can
properly expose below-freezing temperatures. In order to get those,
we need to do an arithmetic right shift with the high-order nibble
in the high byte.

This also stashes the raw packet for later retrieval to make debugging
issues like this easier.